capital_bites_server/requirements.txt
Flask_RESTful==0.3.7
newsdatascraper==0.0.9.1
tensorflow==1.14.0
gunicorn==19.9.0
numpy==1.17.3
Flask==1.1.1
pymongo==3.9.0
typing==3.7.4.1
requests==2.20.0
pandas==0.25.3
Keras==2.2.5
protobuf==3.10.0
google-cloud-language==1.3.0
google-oauth==1.0.1
dnspython==1.16.0
twilio==6.33.0
